+++ CHANGELOG.md
@@ -12,11 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add streaming generate in `api` 
 - Add streaming generate in `message controller` with simple animate
 - Add useStream field in `local`
+- Add useStream option in `setting` page
 
 ### Changed 
 
 - Move all init statement in `main` to `init` file
 - Move the order of fields in `local`
+- Set useless `obs` variable to `var` in `setting controller`
 - Update README
 
 ## [0.1.0] - 2024-07-22